Remove conflicted code + restructure
