--- indexer/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "indexer-substrate"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-regex = "1.5.4"
-serde = { version = "1.0.126", features = ["derive", "rc"] }
-serde_json = "1.0.40"
-postgres = "0.19.1"
-tokio = { version = "0.3.0", features = ["full"] }
-tokio-compat-02 = "0.1.0"
-reqwest = "0.10.8"
