+++ indexer/index-manager/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-jsonrpc-http-server = "15.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.8"
 serde_json = "1.0.40"
@@ -14,6 +13,7 @@ tokio = { version = "0.3.0", features = ["full"] }
 tokio-compat-02 = "0.1.0"
 slog = "2.5.2"
 anyhow = "1.0"
-
+jsonrpc-http-server = "15.0"
 # Massbit dependencies
-#indexer-substrate = { path = "../indexer" }
+json-rpc-server = { path = "../json-rpc-server" }
+ipfs-client = { path = "../ipfs-client" }
