+++ indexer/ipfs-client/Cargo.toml
@@ -12,7 +12,7 @@ http = "0.2.3"
 futures03 = { version = "0.3.1", package = "futures", features = ["compat"] }
 reqwest = { version = "0.10", features = ["json", "stream"] }
 serde = { version = "1.0.126", features = ["rc", "derive"] }
-serde_json = { version = "1.0", features = ["arbitrary_precision"] }
+#serde_json = { version = "1.0", features = ["arbitrary_precision"] }
 serde_yaml = "0.8"
 tokio = { version = "0.2.25", features = ["stream", "rt-threaded", "rt-util", "blocking", "time", "sync", "macros", "test-util", "net"] }
 tokio-stream = { version = "0.1.6", features = ["sync"] }
