+++ indexer/json-rpc-server/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod json_rpc_server;
