--- indexer/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod indexer_substrate;
