--- massbit-indexer/.env
@@ -1,1 +1,0 @@
-DATABASE_URL=postgres://graph-node:let-me-in@localhost/graph-node
