--- massbit-indexer/Cargo.toml
@@ -1,120 +1,0 @@
-[package]
-name = "massbit-indexer"
-version = "0.1.0"
-edition = "2018"
-authors = ["huy <anhhuy0501@gmail.com>"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-dotenv          = "0.15.0"
-env_logger      = { version = "0.7", optional = false }
-ws              = { version = "0.9.1", optional = true, features = ["ssl"] }
-hex             = { version = "0.4.2", default-features = false }
-log             = { version = "0.4", optional = true }
-serde           = { version = "1.0", optional = false }
-serde_json      = { version = "1.0", optional = false }
-thiserror       = { version = "1.0", optional =true }
-primitive-types = { version = "0.6.1", optional = true, features = ["codec"] }
-
-substrate-api-client = { version = "0.6.0",path = "../../substrate-api-client"}
-
-[dependencies.sp-core]
-git = "https://github.com/paritytech/substrate.git"
-branch = "master"
-package = "sp-core"
-default-features = false
-features = ["full_crypto"]
-
-#[dependencies.codec]
-#package = "parity-scale-codec"
-#features = ["derive"]
-#version = "2.0.0"
-#default-features = false
-
-#[dependencies.metadata]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "frame-metadata"
-#default-features=false
-
-#[dependencies.sp-version]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "sp-version"
-#optional = true
-
-#[dependencies.balances]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "pallet-balances"
-#optional = true
-
-#[dependencies.system]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "frame-system"
-#optional = true
-
-#[dependencies.sp-runtime]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "sp-runtime"
-#default-features=false
-
-#[dependencies.support]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "frame-support"
-#default-features=false
-
-#[dependencies.sp-std]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "sp-std"
-#default-features = false
-
-[dependencies.sc-rpc-api]
-git = "https://github.com/paritytech/substrate.git"
-branch = "master"
-optional = true
-
-[dependencies.node-template-runtime]
-git = "https://github.com/scs/substrate-api-client-test-node"
-branch = "bump-to-polkadot-v0.9.2"
-
-#[dev-dependencies.keyring]
-#git = "https://github.com/paritytech/substrate.git"
-#branch = "master"
-#package = "sp-keyring"
-
-[dependencies.clap]
-version = "2.33"
-features = ["yaml"]
-
-[features]
-default = ["std"]
-std = [
-#    "sc-rpc-api",
-    "sp-core/std",
-#    "codec/std",
-#    "metadata/std",
-#    "sp-version",
-#    "balances",
-#    "system",
-#    "sp-runtime/std",
-#    "support/std",
-#    "sp-std/std",
-#    "serde/std",
-#    "serde_json",
-#    "env_logger",
-#    "log",
-#    "ws",
-#    "hex/std",
-#    "primitive-types",
-#    "thiserror",
-]
-# need to add this for the app_crypto macro
-full_crypto = []
-
