--- massbit-indexer/rust-toolchain.bak
@@ -1,1 +1,0 @@
-nightly-2020-10-01
