--- massbit-indexer/src/cli.yml
@@ -1,20 +1,0 @@
-name: "substrate-api-client-examples"
-version: "1.0.0"
-about: Example codes showing the api's capabilities
-authors: "Supercomputing Systems AG <info@scs.ch>"
-
-# AppSettings can be defined as a list and are **not** ascii case sensitive
-settings:
-  - ColoredHelp
-
-args:
-  - node-server:
-      short: u
-      long: url
-      help: Set the node server IP address including ws:// or wss:// prefix
-      takes_value: true
-  - node-port:
-      short: p
-      long: port
-      help: Set the websocket port to listen for substrate events
-      takes_value: true
