--- massbit-indexer/src/lib.rs
@@ -1,12 +1,0 @@
-#[macro_use]
-extern crate clap;
-
-pub mod substrate_chain;
-
-
-// #[macro_use]
-// extern crate dotenv;
-
-// use dotenv::dotenv;
-//use std::env;
-
