--- massbit-indexer/src/main.rs
@@ -1,26 +1,0 @@
-/*
-    Copyright 2019 Supercomputing Systems AG
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
-///! Very simple examples that shows how to pretty print the metadata. Has proven to be a helpful
-///! debugging tool.
-
-use massbit_indexer::substrate_chain;
-
-
-fn main() {
-
-    substrate_chain::get_data();
-}
-
